Remove a bunch of stuff from root_val
